use scout prefix when deleting indexes

+++ src/Engines/MeilisearchEngine.php
@@ -3,6 +3,7 @@
 namespace Laravel\Scout\Engines;
 
 use BackedEnum;
+use Illuminate\Support\Facades\Config;
 use Illuminate\Support\LazyCollection;
 use Laravel\Scout\Builder;
 use Laravel\Scout\Jobs\RemoveableScoutCollection;
@@ -430,6 +431,10 @@ class MeilisearchEngine extends Engine
         $indexes = $this->meilisearch->getIndexes($query);
 
         foreach ($indexes->getResults() as $index) {
+            if (! str($index->getUid())->startsWith(Config::get('scout.prefix'))) {
+                continue;
+            }
+
             $tasks[] = $index->delete();
         }
